Experimenting with new formulation types.
