+++ .gitignore
@@ -10,3 +10,4 @@ docs/site/
 docs/.documenter
 .project
 .tags
+scripts/*.jl
