+++ test/runtests.jl
@@ -54,7 +54,7 @@ include("common.jl")
 
     include("owf.jl")
 
-    #include("des.jl")
+    ##include("des.jl")
 
     include("relax.jl")
